User Interaction Backend/requirements.txt

fastapi
uvicorn[standard]


openai>=1.3.0


chromadb>=0.4.15


sqlalchemy>=2.0
asyncpg 


pydantic>=2.0


python-dotenv


tqdm
rich


orjson


langchain
langgraph

psycopg2
fuzzywuzzy
python-Levenshtein


colorlog